Album Description
First Commit : project description
 The exercise involves building a “proof of concept” app which consumes a REST service and displays photos with
headings and descriptions.

The feed contains a title and a list of rows
The title in the toolbar should be updated from the json 􀀀
 Displays the content (including image, title and description) as a
list of rows 􀀀
Each row should be the right height to display its own content
and no taller. No content should be clipped.This means some
rows will be larger than others.
Loads the images lazily
Don’t download them all at once, but only as needed
Refresh function
either place a refresh button or use a
pull-to-refresh